@@ -1,12 +1,10 @@
-﻿using Zambon.Core.Database.Entity;
-using Zambon.Core.Database.ExtensionMethods;
-using Zambon.Core.Module.Services;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Dynamic.Core;
-using System.Text;
 using System.Text.RegularExpressions;
+using Zambon.Core.Database.ExtensionMethods;
+using Zambon.Core.Module.Interfaces;
+using Zambon.Core.Module.Services;
 
 namespace Zambon.Core.Module.Expressions
 {
@@ -15,111 +13,101 @@ namespace Zambon.Core.Module.Expressions
 
         #region Variables
 
-        //private readonly ICurrentUserService _currentUser;
-
-        private readonly IDictionary<string, object> GlobalParameters = new Dictionary<string, object>();
+        private readonly ICurrentUserService currentUserService;
 
         #endregion
 
         #region Constructors
 
-        public GlobalExpressionsService(ICurrentUserService currentUserService)
+        public GlobalExpressionsService(ICurrentUserService _currentUserService)
         {
-            Insert("_CurrentUserService", currentUserService);
+            currentUserService = _currentUserService;
         }
 
         #endregion
 
         #region Methods
 
-        private void Insert<T>(string key, T _obj) where T : class
+        private object GetReplacementValue(string text)
         {
-            if (!GlobalParameters.ContainsKey(key))
-                GlobalParameters.Add(key, _obj);
-            else
-                GlobalParameters[key] = _obj;
-        }
-
-        public object Get(string key)
-        {
-            if (GlobalParameters.ContainsKey(key))
-                return GlobalParameters[key];
-            return null;
-        }
-        public T Get<T>(string key) where T : class
-        {
-            return (T)Get(key);
-        }
-
-
-        private object GetReplacementValue<T>(string _text, T _obj) where T : class
-        {
-            return (_text.StartsWith("_"))
-                ? Get(_text)
-                : (_obj != null ? (new[] { _obj }).AsQueryable().Select(_text).FirstOrDefault() : null);
-        }
-
-        public string FormatExpressionValue(string _exp, object _obj)
-        {
-            var method = typeof(GlobalExpressionsService).GetMethod("FormatExpressionTypedValue").MakeGenericMethod(_obj.GetType().GetCorrectType());
-            return method.Invoke(this, new object[] { _exp, _obj }).ToString();
-        }
-        public string FormatExpressionTypedValue<T>(string _exp, T _obj) where T : class
-        {
-            if (_exp != null)
-                return Regex.Replace(
-                        _exp,
-                        @"\[(.*?)\]",
-                        m => GetReplacementValue(m.Groups[1].ToString(), _obj).ToString()
-                    );
-            return null;
-        }
-        public object[] FormatExpressionValues(string[] _exp, object _obj = null)
-        {
-            if (_exp != null)
+            switch (text)
             {
-                var items = _exp.Select(x =>
+                case "CurrentUser":
+                    return currentUserService.CurrentUser;
+                default:
+                    return text;
+            }
+        }
+        private object GetReplacementValue<T>(string text, T obj)
+        {
+            switch (text)
+            {
+                case "CurrentUser":
+                    return currentUserService.CurrentUser;
+                default:
+                    return (new[] { obj }).AsQueryable().Select(text).FirstOrDefault();
+            }
+        }
+
+        public object[] FormatArguments(string criteriaArguments)
+        {
+            var arrayArguments = criteriaArguments.Split(",");
+            if ((arrayArguments?.Length ?? 0 ) > 0)
+                return arrayArguments.Select(x =>
                     Regex.Match(x, @"\[(.*?)\]").Success
-                        ? GetReplacementValue(Regex.Match(x, @"\[(.*?)\]").Groups[1].Value, _obj)
+                        ? GetReplacementValue(Regex.Match(x, @"\[(.*?)\]").Groups[1].Value)
                         : x
                 ).ToArray();
-                if (items.Length > 0)
-                    return items;
-            }
-            return new object[] { };
+            return new object[0];
+        }
+        public object[] FormatArguments<T>(string criteriaArguments, T obj)
+        {
+            var arrayArguments = criteriaArguments.Split(",");
+            if ((arrayArguments?.Length ?? 0) > 0)
+                return arrayArguments.Select(x =>
+                    Regex.Match(x, @"\[(.*?)\]").Success
+                        ? GetReplacementValue(Regex.Match(x, @"\[(.*?)\]").Groups[1].Value, obj)
+                        : x
+                ).ToArray();
+            return new object[0];
         }
 
-        public IEnumerable<IExpression> GetApplicableExpressionsItems(IExpression[] expressions, object _obj)
+        public bool IsConditionApplicable(ICondition element, object obj = null)
         {
-            return (IEnumerable<IExpression>)typeof(GlobalExpressionsService).GetMethods().FirstOrDefault(x => x.Name == "GetApplicableExpressionsItems" && x.GetGenericArguments().Count() == 1).MakeGenericMethod(_obj.GetType()).Invoke(this, new[] { expressions, _obj });
+            if (obj == null)
+                return (new[] { new object() }).AsQueryable().Any(element.Condition, FormatArguments(element.ConditionArguments));
+
+            return (bool)GetType().GetMethods().FirstOrDefault(x => x.Name == nameof(IsConditionApplicable) && x.IsGenericMethod).MakeGenericMethod(obj.GetType()).Invoke(this, new[] { element, obj });
         }
-        public IEnumerable<IExpression> GetApplicableExpressionsItems<T>(IExpression[] expressions, T _obj) where T : class
+        public bool IsConditionApplicable<T>(ICondition element, T obj) where T : class
         {
-            var check = (new[] { _obj }).AsQueryable();
-            return expressions.Where(x =>
-                check.Where(
-                    x.ConditionExpression,
-                    FormatExpressionValues(x.ConditionArgumentsList, _obj)).Count() > 0
-            );
+            return (new[] { obj }).AsQueryable().Any(element.Condition, FormatArguments(element.ConditionArguments, obj));
         }
 
-        public bool IsApplicableItem(IExpression expression)
+        public string FormatActionParameters(IActionParameters element, object obj)
         {
-            return (new[] { new object() }).AsQueryable().Count(
-                    expression.ConditionExpression,
-                    FormatExpressionValues(expression.ConditionArgumentsList)) > 0;
-        }
-        public bool IsApplicableItem(IExpression expression, object _obj)
-        {
-            return (bool)typeof(GlobalExpressionsService).GetMethods().FirstOrDefault(x => x.Name == "IsApplicableItem" && x.GetGenericArguments().Count() == 1 && x.GetParameters().Count() == 2).MakeGenericMethod(_obj?.GetType() ?? typeof(object)).Invoke(this, new[] { expression, _obj ?? new object() });
-        }
-        public bool IsApplicableItem<T>(IExpression expression, T _obj) where T : class
-        {
-            return (new[] { _obj }).AsQueryable().Where(
-                    expression.ConditionExpression,
-                    FormatExpressionValues(expression.ConditionArgumentsList, _obj)).Count() > 0;
+            var method = GetType().GetMethods().FirstOrDefault(x => x.Name == nameof(GetReplacementValue) && x.IsGenericMethod).MakeGenericMethod(obj.GetType().GetCorrectType());
+
+            var parameters = element.ActionParameters.Split("&").ToDictionary(k => k.Split("=")[0], v => v.Split("=")[1]);
+            if ((parameters?.Count() ?? 0) > 0)
+                return string.Join("&", parameters.Select(x =>
+                    x.Key + "=" +
+                    (Regex.Match(x.Value, @"\[(.*?)\]").Success
+                        ? method.Invoke(this, new object[] { Regex.Match(x.Value, @"\[(.*?)\]").Groups[1].Value, obj })
+                        : x)
+                ));
+            return string.Empty;
         }
 
+        public IEnumerable<TCondition> GetApplicableItems<TCondition>(TCondition[] conditions, object obj) where TCondition : class, ICondition
+        {
+            return (IEnumerable<TCondition>)GetType().GetMethods().FirstOrDefault(x => x.Name == nameof(GetApplicableItems) && x.IsGenericMethod).MakeGenericMethod(typeof(TCondition), obj.GetType()).Invoke(this, new[] { conditions, obj });
+        }
+        public IEnumerable<TCondition> GetApplicableItems<TCondition, TObject>(TCondition[] expressions, TObject obj) where TCondition : class, ICondition where TObject : class
+        {
+            return expressions.Where(x => IsConditionApplicable(x, obj));
+        }
+        
         #endregion
 
     }
