@@ -1,8 +1,4 @@
-﻿using Zambon.Core.Module.Xml;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Xml.Serialization;
+﻿using System.Xml.Serialization;
 
 namespace Zambon.Core.Module.Xml.Views.DetailViews.Scripts
 {
