@@ -1,10 +1,6 @@
-﻿using Zambon.Core.Module.Operations;
+﻿using System.Xml.Serialization;
 using Zambon.Core.Module.Xml.Views.DetailViews;
 using Zambon.Core.Module.Xml.Views.ListViews;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Xml.Serialization;
 
 namespace Zambon.Core.Module.Xml.Views
 {
